@@ -1,11 +1,12 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./src/**/*.{js,jsx,ts,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
+export const content = [
+  "./index.html",
+  "./src/**/*.{js,jsx,ts,tsx}",
+];
+export const theme = {
+  fontFamily: {
+    'sans': ['ui-sans-serif', 'system-ui']
+  }
+};
+export const plugins = [];
 
